Refractured the update component
